Practice 4: Code Separation and utilising utilities and constants
