+++ src/constants/httpsStatusCodes.js
@@ -1,0 +1,8 @@
+const httpStatusCodes = {
+  OK: 200,
+  Created: 201,
+  NotFound: 404,
+  BadRequest: 400,
+}
+
+module.exports = httpStatusCodes;
