+++ src/modules/products/constants/messages.js
@@ -1,0 +1,8 @@
+const productResponseMessages = {
+  getSuccess: 'Products Fetched Successfully',
+  getError: 'Error fetching Products',
+  createSuccess: 'Product created successfully',
+  createError: 'Error creating Products',
+}
+
+module.exports = productResponseMessages;
