+++ src/utils/response.js
@@ -1,0 +1,16 @@
+const successResponse = (message, data, statusCode = 200) => ({
+    statusCode: statusCode,
+    message,
+    data,
+  });
+
+
+const errorResponse = (message, data, statusCode = 404) => ({
+  statusCode: statusCode,
+  message,
+  data,
+});
+
+module.exports = {
+  successResponse, errorResponse
+}
